+++ airflow/ui/src/components/FlexibleForm/FieldDropdown.tsx
@@ -20,9 +20,9 @@ import { createListCollection } from "@chakra-ui/react/collection";
 import { useRef } from "react";
 
 import { Select } from "src/components/ui";
+import { paramPlaceholder, useParamStore } from "src/queries/useParamStore";
 
 import type { FlexibleFormElementProps } from ".";
-import { paramPlaceholder, useParamStore } from "../../queries/useParamStore";
 
 const labelLookup = (key: string, valuesDisplay: Record<string, string> | undefined): string => {
   if (valuesDisplay && typeof valuesDisplay === "object") {
