+++ airflow/ui/src/components/FlexibleForm/FieldMultiSelect.tsx
@@ -19,8 +19,9 @@
 import { type MultiValue, Select as ReactSelect } from "chakra-react-select";
 import { useState } from "react";
 
+import { paramPlaceholder, useParamStore } from "src/queries/useParamStore";
+
 import type { FlexibleFormElementProps } from ".";
-import { paramPlaceholder, useParamStore } from "../../queries/useParamStore";
 
 const labelLookup = (key: string, valuesDisplay: Record<string, string> | undefined): string => {
   if (valuesDisplay && typeof valuesDisplay === "object") {
