+++ airflow/ui/src/components/FlexibleForm/FieldSelector.tsx
@@ -17,9 +17,9 @@
  * under the License.
  */
 import type { ParamSchema, ParamSpec } from "src/queries/useDagParams";
+import { paramPlaceholder, useParamStore } from "src/queries/useParamStore";
 
 import type { FlexibleFormElementProps } from ".";
-import { paramPlaceholder, useParamStore } from "../../queries/useParamStore";
 import { FieldAdvancedArray } from "./FieldAdvancedArray";
 import { FieldBool } from "./FieldBool";
 import { FieldDateTime } from "./FieldDateTime";
