+++ airflow/ui/src/components/FlexibleForm/FieldStringArray.tsx
@@ -18,8 +18,9 @@
  */
 import { Textarea } from "@chakra-ui/react";
 
+import { paramPlaceholder, useParamStore } from "src/queries/useParamStore";
+
 import type { FlexibleFormElementProps } from ".";
-import { paramPlaceholder, useParamStore } from "../../queries/useParamStore";
 
 export const FieldStringArray = ({ name }: FlexibleFormElementProps) => {
   const { paramsDict, setParamsDict } = useParamStore();
