+++ airflow/ui/src/pages/Connections/AddConnectionButton.tsx
@@ -16,24 +16,39 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { Box, Heading, VStack } from "@chakra-ui/react";
+import { Box, Heading, Spinner, VStack } from "@chakra-ui/react";
 import { useDisclosure } from "@chakra-ui/react";
 import { FiPlusCircle } from "react-icons/fi";
 
 import { Dialog } from "src/components/ui";
 import ActionButton from "src/components/ui/ActionButton";
+import { useConnectionTypeMeta } from "src/queries/useConnectionTypeMeta";
 
 import ConnectionForm from "./ConnectionForm";
 
+export type AddConnectionParams = {
+  conf: string;
+  conn_type: string;
+  connection_id: string;
+  description: string;
+  host: string;
+  login: string;
+  password: string;
+  port: string;
+  schema: string;
+};
+
 const AddConnectionButton = () => {
   const { onClose, onOpen, open } = useDisclosure();
+  const { formattedData: connectionTypeMeta, isPending, keysList: connectionTypes } = useConnectionTypeMeta();
 
   return (
     <Box>
       <ActionButton
         actionName="Add Connection"
         colorPalette="blue"
-        icon={<FiPlusCircle />}
+        disabled={isPending}
+        icon={isPending ? <Spinner size="sm" /> : <FiPlusCircle />}
         onClick={onOpen}
         text="Add Connection"
         variant="solid"
@@ -50,7 +65,11 @@ const AddConnectionButton = () => {
           <Dialog.CloseTrigger />
 
           <Dialog.Body>
-            <ConnectionForm onClose={onClose} />
+            <ConnectionForm
+              connectionTypeMeta={connectionTypeMeta}
+              connectionTypes={connectionTypes}
+              onClose={onClose}
+            />
           </Dialog.Body>
         </Dialog.Content>
       </Dialog.Root>
